+++ src/server/prompts/emailTemplate.js
